% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.nna}
\alias{is.nna}
\title{Function checking if object is not-NA}
\usage{
is.nna(x)
}
\arguments{
\item{x}{Vector of input values}
}
\value{
A logical vector
}
\description{
Function checking if object is not-NA
}
