% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mig.R
\docType{data}
\name{mig_data}
\alias{mig_data}
\title{Eurostat migration data}
\format{
Each is a tibble with two variables and three observations
}
\usage{
mig_data
}
\description{
These data sets have been downloaded from:
the publicly available Eurostat data base on population:
https://ec.europa.eu/eurostat/web/main/data/database,
data bases 'migr' and 'demo'.
These are migration statistics for selected EU countries,
and population sizes. The data relate to 2010-2019 period.
}
\details{
Shiny app made by Maciej Jan Dańko:
https://maciej-jan-danko.shinyapps.io/HMigD_Shiny_App_I/,
These data contain the assumed accuracy of the data sources.

Shiny app made by Maciej Jan Dańko:
https://maciej-jan-danko.shinyapps.io/undercountmigscores/
These data contain undercounting scores for emigration
and immigration based on 30\% metadata and 70\% model scores.

\code{mig_data} are a data frame/tibble.
}
\examples{
mig_data
}
\keyword{datasets}
