% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigMod-package.R
\docType{package}
\name{MigMod-package}
\alias{MigMod-package}
\alias{MigMod}
\title{The 'MigMod' package.}
\description{
ESTIMATES BILATERAL MIGRATION FLOWS FOR SELECTED EU COUNTRIES
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
