% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migmod_m.R
\name{migmod_m}
\alias{migmod_m}
\title{Function to estimate migration model}
\usage{
migmod_m(
  df,
  sending = c("SE", "FI", "IT", "PL"),
  receiving = c("SE", "FI", "IT", "PL"),
  years = 2010:2019,
  ref.country = "SE",
  ...
)
}
\arguments{
\item{df}{A data frame with migration data for sending countries, receiving countries, data quality measures and covariates}

\item{sending}{A vector of ISO2 codes with sending countries to be used in the model (recommended to keep SE and FI)}

\item{receiving}{A vector of ISO2 codes with receiving countries to be used in the model (recommended to keep SE and FI)}

\item{years}{A vector of years to be used in the model (2010 to 2019)}

\item{...}{Arguments passed to function \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} from package \pkg{\link{rstan}} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Function to estimate migration model
}
\seealso{
\code{\link[rstan]{stan}}
}
