% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2_standata.R
\name{data_2_standata}
\alias{data_2_standata}
\title{Function transforming data frame to a list - input to Stan}
\usage{
data_2_standata(
  df,
  sending = c("SE", "FI", "IT", "PL"),
  receiving = c("SE", "FI", "IT", "PL"),
  years = 2010:2019,
  ref.country = "SE"
)
}
\arguments{
\item{df}{A data frame with migration data for sending countries, receiving countries, data quality measures and covariates}

\item{sending}{A vector of ISO2 codes with sending countries to be used in the model (recommended to keep SE and FI)}

\item{receiving}{A vector of ISO2 codes with receiving countries to be used in the model (recommended to keep SE and FI)}

\item{years}{A vector of years to be used in the model (2010 to 2019)}

\item{ref.country}{A character with an ISO2 code receiving country which immigration data are is used as a reference. This assumes that the data reported by this country are not biased. This assumption ensures identification of model parameters. It is strongly recommended to use Sweden ("SE") as a reference.}
}
\value{
A list to be passed on to the stan model that contains all required inputs
}
\description{
Function transforming data frame to a list - input to Stan
}
