% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_2_standata}
\alias{data_2_standata}
\title{Function transforming data frame with data to standata (a list)}
\usage{
data_2_standata(
  df,
  sending = c("SE", "FI", "IT", "PL"),
  receiving = c("SE", "FI", "IT", "PL"),
  years = 2010:2019
)
}
\arguments{
\item{df}{A data frame with migration data for sending countries, receiving countries, data quality measures and covariates}

\item{sending}{A vector of ISO2 codes with sending countries to be used in the model (recommended to keep SE and FI)}

\item{receiving}{A vector of ISO2 codes with receiving countries to be used in the model (recommended to keep SE and FI)}

\item{years}{A vector of years to be used in the model (2010 to 2019)}
}
\value{
A list to be passed on to the stan model that contains all required inputs
}
\description{
Function transforming data frame with data to standata (a list)
}
